cleaning(core): move PrsimicSharedConfig to node-utils
