--- packages/core/src/filesystem/index.ts
@@ -1,1 +1,0 @@
-export * from "./PrismicSharedConfig";
