+++ packages/core/src/index.ts
@@ -1,4 +1,3 @@
-export * as FileSystem from "./filesystem";
 export * as Utils from "./utils";
 export * from "./core";
 export * as Models from "./models";
