+++ packages/core/src/node-utils/index.ts
@@ -2,3 +2,4 @@ export * as Framework from "./framework";
 export * from "./manifest";
 export * from "./paths";
 export * from "./pkg";
+export * from "./PrismicSharedConfig";
