+++ packages/slice-machine/server/src/api/auth/start.ts
@@ -1,4 +1,4 @@
-import { FileSystem, Utils } from "@slicemachine/core";
+import { NodeUtils, Utils } from "@slicemachine/core";
 
 type StartResponse = {
   err?: Error;
@@ -8,7 +8,7 @@ type StartResponse = {
 export default async function handler(): Promise<StartResponse> {
   try {
     // Reset the prismic auth cookie
-    FileSystem.PrismicSharedConfigManager.setProperties({
+    NodeUtils.PrismicSharedConfigManager.setProperties({
       cookies: Utils.Cookie.serializeCookies([]),
     });
     return {};
